pyproject.toml
[project]
name = "myaa"
version = "0.1.0"
description = "Minimal LangGraph-powered Discord assistant"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.11"

dependencies = [
    "langgraph>=0.4.3",
    "discord.py>=2.5.2",
    "python-dotenv>=1.1.0",
    "openai>=1.78.1",
    "tqdm>=4.66.5"
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "ruff>=0.11.8",
    "pytest>=8.3.5",
]

[tool.uv]
index-url = "https://pypi.org/simple"

[project.scripts]
run-discord = "myaa.adapter.discord.run:entrypoint"